// libs/features/medical-board/src/lib/shared/models/mb-details.ts
import { BilingualText } from '@gosi-ui/core';
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
export class MbDetails {
  medicalBoardType: BilingualText = new BilingualText();
  officeLocation: BilingualText = new BilingualText();
  sessionChannel: BilingualText = new BilingualText();
}